Pulling common functions into more spectific util files.
#1493
