+++ waltz-ng/client/entity-statistics/components/entity-statistic-detail-table.js
@@ -17,7 +17,8 @@
  */
 import _ from "lodash";
 import moment from "moment";
-import {notEmpty, mkEntityLinkGridCell, mkLinkGridCell} from "../../common";
+import {notEmpty} from "../../common";
+import {mkEntityLinkGridCell, mkLinkGridCell} from "../../common/link-utils";
 
 
 /**
