+++ waltz-ng/client/logical-flow/components/logical-flow-table/logical-flow-table.js
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 import _ from "lodash";
-import {mkEntityLinkGridCell} from "../../../common";
+import {mkEntityLinkGridCell} from "../../../common/link-utils";
 
 
 const bindings = {
