+++ waltz-ng/client/measurable-rating/measurable-rating-edit.js
@@ -17,7 +17,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 import _ from 'lodash';
-import {initialiseData, kindToViewState} from '../common';
+import {initialiseData} from '../common';
+import {kindToViewState} from '../common/link-utils';
 
 
 const initialState = {
